# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: MOVE Apayao — Transit & Trip Booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get current user's wallet
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletResponse"
  /wallet/transactions:
    get:
      operationId: listTransactions
      tags: [wallet]
      summary: Get transaction history
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTransactionsResponse"
  /wallet/topup:
    post:
      operationId: topUpWallet
      tags: [wallet]
      summary: Admin-only direct wallet credit (cash kiosk, manual adjustment/refund)
      description: >
        Directly credits a wallet with no payment-provider step. Restricted to
        admins. Passengers topping up their own wallet must use
        /wallet/topup/intent, which requires a verified payment before the
        ledger is touched.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopUpBody"
      responses:
        "201":
          description: Top-up successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopUpResponse"
  /wallet/topup/intent:
    post:
      operationId: createTopUpIntent
      tags: [wallet]
      summary: Create a wallet top-up checkout session (passenger-callable)
      description: >
        Does not credit the wallet. Creates a pending top-up and returns a
        payment-provider checkout URL; the wallet is only credited once the
        provider confirms payment via /wallet/topup/webhook.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopUpIntentBody"
      responses:
        "201":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopUpIntentResponse"
  /wallet/tap:
    post:
      operationId: tapWallet
      tags: [wallet]
      summary: NFC/QR card tap — deduct fare (mock reader interface)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TapBody"
      responses:
        "200":
          description: Tap processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TapResponse"
        "402":
          description: Insufficient balance
  /wallet/admin/users:
    get:
      operationId: listWallets
      tags: [wallet]
      summary: List all user wallets (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Wallet list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWalletsResponse"

  # ── Incidents ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Wallet:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
          format: uuid
        balance:
          type: number
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, balance, updatedAt]
    FareTransaction:
      type: object
      properties:
        id:
          type: integer
        walletId:
          type: integer
        userId:
          type: string
          format: uuid
        amount:
          type: number
        type:
          type: string
          enum: [top_up, deduct, refund, adjustment]
        description:
          type: string
          nullable: true
        referenceId:
          type: string
          nullable: true
        balanceBefore:
          type: number
        balanceAfter:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, walletId, userId, amount, type, balanceBefore, balanceAfter, createdAt]
    WalletResponse:
      type: object
      properties:
        wallet:
          $ref: "#/components/schemas/Wallet"
      required: [wallet]
    ListTransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/FareTransaction"
        limit:
          type: integer
        offset:
          type: integer
      required: [transactions, limit, offset]
    TopUpBody:
      type: object
      description: Admin-only direct credit.
      properties:
        userId:
          type: string
          format: uuid
        amount:
          type: number
        referenceId:
          type: string
      required: [userId, amount]
    TopUpResponse:
      type: object
      properties:
        transaction:
          $ref: "#/components/schemas/FareTransaction"
        newBalance:
          type: number
      required: [transaction, newBalance]
    TopUpIntentBody:
      type: object
      description: Passenger-callable. Does not credit the wallet directly.
      properties:
        amount:
          type: number
          maximum: 50000
      required: [amount]
    TopUpIntentResponse:
      type: object
      properties:
        checkoutUrl:
          type: string
      required: [checkoutUrl]
    TapBody:
      type: object
      properties:
        cardId:
          type: string
        amount:
          type: number
        tripId:
          type: integer
        idempotencyKey:
          type: string
          description: Reader-generated dedup key so a network retry can't double-deduct.
      required: [cardId, amount, idempotencyKey]
    TapResponse:
      type: object
      properties:
        ok:
          type: boolean
        newBalance:
          type: number
        transactionId:
          type: integer
      required: [ok, newBalance, transactionId]
    ListWalletsResponse:
      type: object
      properties:
        wallets:
          type: array
          items:
            $ref: "#/components/schemas/Wallet"
        limit:
          type: integer
        offset:
          type: integer
      required: [wallets, limit, offset]

    # ── Incidents ────────────────────────────────────────────────────────────
...